HDDS-12039. Enhancements to ozone repair om commands
