+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/repair/om/OMRepair.java
@@ -19,6 +19,7 @@
 package org.apache.hadoop.ozone.repair.om;
 
 import org.apache.hadoop.hdds.cli.RepairSubcommand;
+import org.apache.hadoop.ozone.repair.om.quota.QuotaRepair;
 import org.kohsuke.MetaInfServices;
 import picocli.CommandLine;
 
@@ -29,7 +30,8 @@ import picocli.CommandLine;
     subcommands = {
         FSORepairTool.class,
         SnapshotRepair.class,
-        TransactionInfoRepair.class
+        TransactionInfoRepair.class,
+        QuotaRepair.class
     },
     description = "Operational tool to repair OM.")
 @MetaInfServices(RepairSubcommand.class)
