+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/repair/om/SnapshotChainRepair.java
@@ -83,6 +83,9 @@ public class SnapshotChainRepair extends RepairTool {
 
   @Override
   public void execute() throws Exception {
+    if (checkIfServiceIsRunning("OM")) {
+      return;
+    }
     List<ColumnFamilyHandle> cfHandleList = new ArrayList<>();
     List<ColumnFamilyDescriptor> cfDescList = RocksDBUtils.getColumnFamilyDescriptors(dbPath);
