+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/repair/om/quota/QuotaRepair.java
@@ -16,11 +16,10 @@
  *  limitations under the License.
  */
 
-package org.apache.hadoop.ozone.repair.quota;
+package org.apache.hadoop.ozone.repair.om.quota;
 
 import java.io.IOException;
 import java.util.Collection;
-import org.apache.hadoop.hdds.cli.RepairSubcommand;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.ipc.ProtobufRpcEngine;
 import org.apache.hadoop.ipc.RPC;
@@ -30,10 +29,8 @@ import org.apache.hadoop.ozone.om.protocolPB.Hadoop3OmTransportFactory;
 import org.apache.hadoop.ozone.om.protocolPB.OmTransport;
 import org.apache.hadoop.ozone.om.protocolPB.OzoneManagerProtocolClientSideTranslatorPB;
 import org.apache.hadoop.ozone.om.protocolPB.OzoneManagerProtocolPB;
-import org.apache.hadoop.ozone.repair.OzoneRepair;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.ratis.protocol.ClientId;
-import org.kohsuke.MetaInfServices;
 import picocli.CommandLine;
 
 import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_ADDRESS_KEY;
@@ -48,18 +45,14 @@ import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_SERVICE_IDS_KEY;
         QuotaTrigger.class,
     },
     description = "Operational tool to repair quota in OM DB.")
-@MetaInfServices(RepairSubcommand.class)
-public class QuotaRepair implements RepairSubcommand {
-
-  @CommandLine.ParentCommand
-  private OzoneRepair parent;
+public class QuotaRepair {
 
   public OzoneManagerProtocolClientSideTranslatorPB createOmClient(
       String omServiceID,
       String omHost,
       boolean forceHA
   ) throws Exception {
-    OzoneConfiguration conf = parent.getOzoneConf();
+    OzoneConfiguration conf = new OzoneConfiguration();
     if (omHost != null && !omHost.isEmpty()) {
       omServiceID = null;
       conf.set(OZONE_OM_ADDRESS_KEY, omHost);
@@ -93,7 +86,7 @@ public class QuotaRepair implements RepairSubcommand {
   }
 
   private Collection<String> getConfiguredServiceIds() {
-    OzoneConfiguration conf = parent.getOzoneConf();
+    OzoneConfiguration conf = new OzoneConfiguration();
     Collection<String> omServiceIds =
         conf.getTrimmedStringCollection(OZONE_OM_SERVICE_IDS_KEY);
     return omServiceIds;
